service_film_/requirements.txt
# system requirements
pip==23.1.2
setuptools==67.8.0
wheel==0.40.0

# project requirements
anyio==3.6.2
elasticsearch[async]==8.7.0
fastapi==0.95.2
fastapi-cache2[redis]==0.2.1
gunicorn==20.1.0
idna==3.4
mypy==1.3.0
orjson==3.8.5
pydantic==1.10.7
redis==4.5.5
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.0
uvicorn==0.22.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"

# need for faker data
Faker==4.0.0
pandas==2.0.2
tqdm==4.65.0